refactor: improve CLI coverage structure
